+++ apps/chat/src/components/NavigationWrapper.tsx
@@ -30,7 +30,7 @@ import { DEFAULT_CONVERSATION_NAME } from '../constants/default-ui-settings';
 import { MarketplaceTabs } from '@/src/constants/marketplace';
 import { Routes } from '@/src/constants/routes';
 
-import { PromptDialogs } from './Promptbar/components/PromptDialogs/PromptDialogs';
+import { PromptDialogs } from './Promptbar/components/PromptDialogs';
 import { Chatbar } from '@/src/components/Chatbar/Chatbar';
 import { ModelIcon, ModelTooltip } from '@/src/components/Chatbar/ModelIcon';
 import Loader from '@/src/components/Common/Loader';
